requirements.txt
django==2.2
django-crispy-forms==1.12.0
pillow==8.3.1
django-import-export==2.5.0
django-filter==2.4.0
pandas==1.3.1
numpy==1.21.1
django-pandas==0.6.4
reportlab==3.6.1
django-ckeditor==6.1.0